fix(dependencies): Move ManualAnalyzer to api
Rationale:
- workflow-steps should only depend on api and model
- the ManualAnalyzer is a small standalone extension of Analyzer in api
