+++ antenna-api/src/main/java/org/eclipse/sw360/antenna/api/workflow/ManualAnalyzer.java
@@ -8,10 +8,9 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.workflow.analyzers;
+package org.eclipse.sw360.antenna.api.workflow;
 
 import org.eclipse.sw360.antenna.api.exceptions.AntennaConfigurationException;
-import org.eclipse.sw360.antenna.api.workflow.AbstractAnalyzer;
 
 import java.io.File;
 import java.nio.file.Path;
